feat(iot-manager): 产品告警规则更新功能完成

+++ pangu-common/pangu-common-zabbix/src/main/java/com/pangu/common/zabbix/api/ZbxTrigger.java
@@ -57,7 +57,7 @@ public interface ZbxTrigger {
     @JsonPath("/trigger/trigger.update")
     String triggerUpdate(@ParamName("triggerId") String triggerId,
                          @ParamName("expression") String expression,
-                         @ParamName("ruleLevel") Byte ruleLevel);
+                         @ParamName("ruleLevel") Integer ruleLevel);
 
 
     /**
